cleanup(scap): move common/falcosecurity/log.h to userspace/libscap/scap_log.h
