+++ userspace/libscap/debug_log_helpers.h
@@ -18,7 +18,7 @@ limitations under the License.
 #ifndef __DEBUG_LOG_HELPERS_H
 #define __DEBUG_LOG_HELPERS_H
 
-#include "falcosecurity/log.h"
+#include "scap_log.h"
 
 #include <stdio.h>
