+++ userspace/libscap/scap_open.h
@@ -28,7 +28,7 @@ limitations under the License.
 #include "scap_limits.h"
 #include "scap_procs.h"
 #include "../../driver/ppm_events_public.h"
-#include "falcosecurity/log.h"
+#include "scap_log.h"
 
 #ifdef __cplusplus
 extern "C"
